app/requirements.txt
streamlit==1.28.0
opencv-python==4.8.0.76
numpy==1.24.3
psycopg2-binary==2.9.6
mlflow==2.6.0
pandas==2.0.3
Pillow==10.0.0